Update and rename rosnp_helpers.py to rosnp_helpers.py
Fixed case where uint8 data type is used. ROS serializes and deserializes the uint8 array as a `bytes` object instead of a tuple, causing problems when using with Numpy array structures. 

Solution: Cast data as tuples before assigning to message and upon receipt.

This was an intense oversight in my planning that took almost two hours to diagnose and fix, but I did it, PY.